Fix error with what page on which Card component appears
I fixed an error with the page my card shows up on. It had been showing up on the products page (/products) rather than the search page I was assigned to. I was a bit confused on what page I was working on at first, but I just noticed this error a little while ago. So my branch currently creates both the product and now a search page. For testing purposes, I linked the search page to the product button (the circle with addition symbol) because I don't think the search bar is hooked up just yet.